requirements.txt
python-docx==1.1.0
dashscope>=1.20.0
PyPDF2>=3.0.0
lxml>=4.9.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
